requirements.txt
awscli
pycryptodome
blinker
boto3
botocore
click
Flask
Flask-Cors
Flask-JWT-Extended
gunicorn
importlib-metadata
itsdangerous
Jinja2
jmespath
MarkupSafe
pyasn1
PyJWT
python-dateutil
python-dotenv
rsa
s3transfer
six
urllib3
Werkzeug
zipp